:hammer: Move cross-platform utility methods to commonMain
